feat: update Event logic

+++ ewm/src/main/java/ru/practicum/event/dto/EventFullDto.java
@@ -18,7 +18,7 @@ public class EventFullDto {
     private final boolean paid;
     private final LocalDateTime eventDate;
     private final UserShortDto initiator;
-    private final Long confirmedRequests;
+    private final Integer confirmedRequests;
     private final Long views;
     private final String description;
     private final Integer participantLimit;
